verification.py
from llama_index.core import VectorStoreIndex, SimpleDirectoryReader
import faiss
import pandas as pd

def verify_labels(df):
    """Use LlamaIndex & FAISS to verify labeled data."""
    index = VectorStoreIndex(faiss.IndexFlatL2(df.shape[1]))
    index.insert(df.to_numpy())
    return df  # Placeholder: Implement verification logic
