requirements.txt
fastapi
uvicorn
pandas
numpy
ollama
autogen
llama-index
faiss-cpu
matplotlib
plotly
scipy
